Fix weird error in copper harvest level Move static field in mixin class to fix ExceptionInInitializerError when using EFR as a library without it being in the mods folder

+++ src/main/java/ganymedes01/etfuturum/blocks/itemblocks/ItemBlockWoodDoor.java
@@ -25,7 +25,8 @@ public class ItemBlockWoodDoor extends ItemBlock {
 				return false;
 			ItemDoor.placeDoorBlock(world, x, y, z, MathHelper.floor_double((player.rotationYaw + 180.0F) * 4.0F / 360.0F - 0.5D) & 3, field_150939_a);
 			//Disable the sound for continuity, so it doesn't play when the event-based player would not
-			if(ConfigSounds.fixSilentPlacing)world.playSoundEffect((double)((float)x + 0.5F), (double)((float)y + 0.5F), (double)((float)z + 0.5F), this.field_150939_a.stepSound.func_150496_b(), (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F, this.field_150939_a.stepSound.getPitch() * 0.8F);
+			if (ConfigSounds.fixSilentPlacing)
+				world.playSoundEffect((float) x + 0.5F, (float) y + 0.5F, (float) z + 0.5F, this.field_150939_a.stepSound.func_150496_b(), (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F, this.field_150939_a.stepSound.getPitch() * 0.8F);
 			stack.stackSize--;
 			return true;
 		}
